+++ R/google_places.R
@@ -10,7 +10,7 @@
 #' @param keyword \code{string} A term to be matched against all content that Google has indexed for this place, including but not limited to name, type, and address, as well as customer reviews and other third-party content.
 #' @param language \code{string} The language code, indicating in which language the results should be returned, if possible. Searches are also biased to the selected language; results in the selected language may be given a higher ranking. See the list of supported languages and their codes \url{https://developers.google.com/maps/faq#languagesupport}
 #' @param name \code{string} \code{vector} One or more terms to be matched against the names of places. Ignored when used with a \code{search_string}. Results will be restricted to those containing the passed \code{name} values. Note that a place may have additional names associated with it, beyond its listed name. The API will try to match the passed name value against all of these names. As a result, places may be returned in the results whose listed names do not match the search term, but whose associated names do.
-#' @param place_type \code{string} Restricts the results to places matching the specified type. Only one type may be specified (if more than one type is provided, all types following the first entry are ignored). For a list of valid types, please visit \url{https://developers.google.com/places/supported_types}
+#' @param place_type \code{string} Restricts the results to places matching the specified type. Only one type may be specified. For a list of valid types, please visit \url{https://developers.google.com/places/supported_types}
 #' @param price_range \code{numeric} \code{vector} Specifying the minimum and maximum price ranges. Values range between 0 (most affordable) and 4 (most expensive)
 #' @param open_now \code{logical} Returns only those places that are open for business at the time the query is sent. Places that do not specify opening hours in the Google Places database will not be returned if you include this parameter in your query.
 #' @param page_token \code{string} Returns the next 20 results from a previously run search. Setting a \code{page_token} parameter will execute a search with the same parameters used in a previous search. All parameters other than \code{page_token} will be ignored. The \code{page_token} can be found in the result set of a previously run query.
@@ -161,6 +161,12 @@ google_places <- function(search_string = NULL,
   if(!is.null(price_range)){
     if(!price_range[1] %in% 0:4 | !price_range[2] %in% 0:4)
       stop("price_range must be between 0 and 4 inclusive")
+  }
+
+  ## check place type
+  if(!is.null(place_type)){
+    if(length(place_type) > 1 | !is.character(place_type))
+      stop("place_type must be a string vector of length 1")
   }
 
   ## open_now is boolean